switch to dtui table
