+++ src/js/components/bulkDownload/archive/table/AwardDataArchiveTable.jsx
@@ -5,37 +5,64 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import { Table } from 'data-transparency-ui';
 import TableRow from './TableRow';
 
-const propTypes = { results: PropTypes.array, columns: PropTypes.array };
+const columns = [
+    {
+        title: 'agency',
+        displayName: 'Agency'
+    },
+    {
+        title: 'fileName',
+        displayName: 'Archive File'
+    },
+    {
+        title: 'fy',
+        displayName: 'Fiscal Year'
+    },
+    {
+        title: 'date',
+        displayName: 'Data As Of'
+    }
+];
 
-const AwardDataArchiveTable = ({ results, columns }) => {
+const propTypes = { results: PropTypes.array };
+
+const AwardDataArchiveTable = ({ results }) => {
     let noResultsClass = '';
 
     // remove duplicated bottom border
     if (results.length === 0) noResultsClass = ' no-results';
 
-    const headers = columns.map((column) => (
-        <th key={column.columnName}>
-            {column.displayName}
-        </th>
-    ));
+    // const rows = results.map((file, index) => (
+    //     <TableRow
+    //         key={file.url}
+    //         file={file}
+    //         rowIndex={index}
+    //         columns={columns} />
+    // ));
 
-    const rows = results.map((file, index) => (
-        <TableRow
-            key={file.url}
-            file={file}
-            rowIndex={index}
-            columns={columns} />
-    ));
+    const rows = results.map((file) => ([
+        file.agency,
+        file.fileName,
+        file.fy,
+        file.date
+    ]))
+
+    console.log({ rows, results })
 
     return (
-        <div className={`award-data-archive-table${noResultsClass}`}>
-            <table>
-                <thead><tr>{headers}</tr></thead>
-                <tbody>{rows}</tbody>
-            </table>
-        </div>
+        // <div className={`award-data-archive-table${noResultsClass}`}>
+        //     <table>
+        //         <thead><tr>{headers}</tr></thead>
+        //         <tbody>{rows}</tbody>
+        //     </table>
+        // </div>
+        <Table
+            classNames={`award-data-archive-table${noResultsClass}`}
+            columns={columns}
+            rows={rows} />
     );
 }
